removing warnings

+++ meshscript/canvas.cpp
@@ -941,7 +941,7 @@ void canvas::render_pointclouds_on_image(const scene* s, const jtk::image<pixel>
         }
       bind(_rd, camera_position, object_system, projection_mat);
       object_buffer ob;
-      ob.number_of_vertices = pc.p_vertices->size();
+      ob.number_of_vertices = (uint32_t)pc.p_vertices->size();
       ob.vertices = (const float*)pc.p_vertices->data();      
       ob.normals = _settings.shading ? (const float*)pc.p_normals->data() : nullptr;
       ob.colors = _settings.one_bit ? nullptr : (const uint32_t*)pc.p_vertex_colors->data();
